% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_day_of_wyear.R
\name{ice_day_of_wyear}
\alias{ice_day_of_wyear}
\title{Calculate the day of the water year}
\usage{
ice_day_of_wyear(data, year_column = "waterYear")
}
\arguments{
\item{data}{Data frame used to calculate the day of water year}
}
\value{
A data frame
}
\description{
Calculate the day of the water year
}
