% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_group_1.R
\name{ice_group_1}
\alias{ice_group_1}
\title{Calculate total continuous ice coverage per year}
\usage{
ice_group_1(data, year_column = "waterYear")
}
\arguments{
\item{data}{data frame used to calculate ice cover}
}
\value{
A data frame.
}
\description{
Calculate total continuous ice coverage per year
}
