% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_group_2.R
\name{ice_group_2}
\alias{ice_group_2}
\title{Calculate freeze and thaw dates}
\usage{
ice_group_2(
  data,
  year_column = "waterYear",
  value_column = "Value",
  symbol_column = "Symbol",
  date_column = "Date",
  day_of_yr_column = "day_of_year"
)
}
\arguments{
\item{data}{data frame used to calculate freeze and thaw dates}
}
\value{
A data frame
}
\description{
Calculate freeze and thaw dates
}
