% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_group_3.R
\name{ice_group_3}
\alias{ice_group_3}
\title{Calculate onset of freshet}
\usage{
ice_group_3(data)
}
\arguments{
\item{data}{input dataframe used to calculate the onset of freshet}
}
\value{
A data frame with the date, flow value and day of the year for the onset of freshet dates for each year
}
\description{
Calculate onset of freshet
}
