% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_remove_yrs.R
\name{ice_remove_yrs}
\alias{ice_remove_yrs}
\title{Remove years with more than 14 consecutive NA values for flow}
\usage{
ice_remove_yrs(
  data,
  value_column = "Value",
  year_column = "waterYear",
  consecutive_nas = 14
)
}
\arguments{
\item{data}{A data frame}
}
\value{
Data frame without years that have 14 or more NA values for flow
}
\description{
Remove years with more than 14 consecutive NA values for flow
}
