% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_day_of_wyear.R
\name{ice_day_of_wyear}
\alias{ice_day_of_wyear}
\title{Calculate the day of the water year}
\usage{
ice_day_of_wyear(data, year_column = "waterYear")
}
\arguments{
\item{year_column}{column of year values in the dataframe}

\item{Data}{data frame used to calculate the day of the year}
}
\value{
A data frame
}
\description{
Calculate the day of the water year
}
