% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_group_2.R
\name{ice_group_2}
\alias{ice_group_2}
\title{Calculate freeze and thaw dates}
\usage{
ice_group_2(
  data,
  year_column = "waterYear",
  value_column = "Value",
  symbol_column = "Symbol",
  date_column = "Date",
  day_of_yr_column = "day_of_year"
)
}
\arguments{
\item{data}{input data used to calculate the freeze and thaw dates}

\item{year_column}{column of year values in the dataframe}

\item{value_column}{column of flow values in the dataframe}

\item{symbol_column}{column of symbol values in the dataframe}

\item{date_column}{column of daily date values in the dataframe}

\item{day_of_yr_column}{column of day of the year in the dataframe}
}
\value{
A data frame with freeze and thaw dates for each year
}
\description{
Calculate freeze and thaw dates
}
