% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_group_3.R
\name{ice_group_3}
\alias{ice_group_3}
\title{Calculate onset of freshet}
\usage{
ice_group_3(data)
}
\arguments{
\item{data}{A data frame used to calculate onset of freshet}
}
\value{
A data frame with the onset of freshet dates, flow values and day of the year.
}
\description{
Calculate onset of freshet
}
