% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_remove_yrs.R
\name{ice_remove_yrs}
\alias{ice_remove_yrs}
\title{Remove years with more than x consecutive NA values for flow}
\usage{
ice_remove_yrs(
  data,
  value_column = "Value",
  year_column = "waterYear",
  consecutive_nas = 14
)
}
\arguments{
\item{data}{A data frame}

\item{value_column}{column of flow values in the dataframe}

\item{year_column}{column of year values in the dataframe}

\item{consecutive_nas}{non-zero integer}
}
\value{
Data frame with the year(s) removed that have x or more consecutive NA values for flow
}
\description{
Remove years with more than x consecutive NA values for flow
}
